fix: implement ToSingleRedisArg for #[derive(ToRedisArgs)]
As far as I can tell, all #[derive(ToRedisArgs)] implementations we
provide serialize into a single redis value, so we can auto-implement
ToSingleRedisArg for them as well.

This is required for most of the types that have ToRedisArgs
implemented, so that they can be used as a single value, or as a key.